Add Workspace Initializer
